+++ src/core/shared/configurations/shared.ts
@@ -5,6 +5,9 @@ export const SHARED_CONFIG = {
     MAX_BLIP_STREAM_DISTANCE: 650,
     // PLAYERS
     FORCE_SCREEN_SHAKE: true, // Used to shake the player's screen. Makes most aimbots a hell of a lot worse.
+    // FOOD & WATER
+    FOOD_FATIGUE: 10, // If food drops below this number the player stops sprinting.
+    WATER_FATIGUE: 8, // If water drops below this number the player stops sprinting.
     // Inventory
     MAX_PICKUP_RANGE: 2,
     // Interaction