+++ documentation/changelog.md
@@ -49,6 +49,10 @@ description: List of all the changes and current versions.
 + Added Task Timelines to Server Side
 + Timelines allow you to call a set of native player tasks to perform actions in order.
 + Upgrade Vehicle Repair
++ If water drops below shared config setting players no longer can run
++ If food drops below shared config setting players no longer can run
++ Add Seatbelt Sound
++ Fix Command Suggestions
 ```
 
 ## 1.5.0 - January 28-31, 2021